Fixed bug in ari float conversion

+++ pytorch/metrics.py
@@ -39,8 +39,8 @@ def adjusted_rand_index(true_mask, pred_mask):
     true_group_ids = torch.argmax(true_mask, -1)
     pred_group_ids = torch.argmax(pred_mask, -1)
     # We convert true and predicted clusters to one-hot ('oh') representations.
-    pred_mask_oh = torch.nn.functional.one_hot(pred_group_ids, n_pred_groups)
-
+    pred_mask_oh = torch.nn.functional.one_hot(pred_group_ids, n_pred_groups).float()
+    true_mask = true_mask.float()
     # Sum along dim 1 and 2 to get the number of points in each cluster.
     n_points = torch.sum(true_mask, dim=(1, 2))
